Fixed ES6 syntax and create additional modules
